requirements.txt

boto3==1.34.162
botocore==1.34.162
boxsdk[jwt]==3.14.0
faiss-cpu==1.12.0
langchain-aws==0.2.7
langchain-community==0.3.29
langchain-text-splitters==0.3.9
langchain==0.3.27
langsmith==0.4.21
pypdf==6.0.0
python-dotenv==1.1.1
streamlit==1.49.1
